% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadem_function.R
\name{loadem}
\alias{loadem}
\title{A loading multiple libraries Function}
\usage{
loadem(libs)
}
\arguments{
\item{The}{libraries you want to upload. No default.}
}
\description{
This function allows you to upload multiple libraries at once
}
\examples{
loadem(c("dplyr","scales"))
}

